Debug name mismatch logistic_regression and LogisticRegression

+++ backend/server.py
@@ -1,8 +1,8 @@
 import os
 import timeit
 from PIL import Image
-
-# 
+from pathlib import Path
+ 
 import torch
 from torchvision import transforms, datasets
 
@@ -116,19 +116,28 @@ def train_and_save(model_type):
     try: 
         start_timer = timeit.default_timer()
         if model_type == "CNN":
-            training_results = CNN(device)
-            CNN_model = training_results["model"]
-            torch.save(CNN_model.state_dict(), "api_tests/test_models/CNN_model.pth")
+            if Path("api_tests/test_models/CNN_model.pth").exists():
+                print("CNN model exists, clear it first if you want to retrain.")
+            else:
+                training_results = CNN(device)
+                CNN_model = training_results["model"]
+                torch.save(CNN_model.state_dict(), "api_tests/test_models/CNN_model.pth")
             
         elif model_type == "FNN":
-            training_results = FNN(device)
-            FNN_model = training_results["model"]
-            torch.save(FNN_model.state_dict(), "api_tests/test_models/FNN_model.pth")
+            if Path("api_tests/test_models/FNN_model.pth").exists():
+                print("FNN model exists, clear it first if you want to retrain.")
+            else:
+                training_results = FNN(device)
+                FNN_model = training_results["model"]
+                torch.save(FNN_model.state_dict(), "api_tests/test_models/FNN_model.pth")
             
         elif model_type == "LR":
-            training_results = LogisticRegression(device)
-            LR_model = training_results["model"]
-            torch.save(LR_model.state_dict(), "api_tests/test_models/LR_model.pth")
+            if Path("api_tests/test_models/LR_model.pth").exists():
+                print("LR model exists, clear it first if you want to retrain.")
+            else:
+                training_results = LogisticRegression(device)
+                LR_model = training_results["model"]
+                torch.save(LR_model.state_dict(), "api_tests/test_models/LR_model.pth")
         else:
             return (
                 jsonify(
